add report each services

+++ layanan/models.py
@@ -13,7 +13,7 @@ class JenisLayanan(models.Model):
     id_jenis = models.AutoField(primary_key=True)
     layanan = models.ForeignKey(Layanan, on_delete=models.CASCADE, related_name="jenis_layanan")
     nama_jenis = models.CharField(max_length=50)  # Contoh: WiFi Paket 115
-    tarif = models.DecimalField(max_digits=10, decimal_places=2)
+    tarif = models.IntegerField()
 
     def __str__(self):
         return f"{self.layanan.nama_layanan} - {self.nama_jenis}"
